Project organisation + main + embed apps.json
